#284: refactoring and completion of sync http client
