+++ modules/cxf-client-rest/src/main/java/com/devonfw/module/cxf/common/impl/client/rest/SyncServiceClientFactoryCxfRest.java
@@ -5,7 +5,6 @@ import java.util.Collection;
 import java.util.List;
 
 import javax.inject.Inject;
-import javax.ws.rs.Path;
 
 import org.apache.cxf.jaxrs.client.Client;
 import org.apache.cxf.jaxrs.client.ClientConfiguration;
@@ -15,7 +14,8 @@ import org.apache.cxf.jaxrs.client.WebClient;
 import com.devonfw.module.cxf.common.impl.client.SyncServiceClientFactoryCxf;
 import com.devonfw.module.service.common.api.client.context.ServiceContext;
 import com.devonfw.module.service.common.api.client.sync.SyncServiceClientFactory;
-import com.devonfw.module.service.common.api.constants.ServiceConstants;
+import com.devonfw.module.service.common.base.client.ServiceClientTypeHandler;
+import com.devonfw.module.service.common.base.client.ServiceClientTypeHandlerRest;
 import com.fasterxml.jackson.jaxrs.json.JacksonJsonProvider;
 
 /**
@@ -33,6 +33,12 @@ public class SyncServiceClientFactoryCxfRest extends SyncServiceClientFactoryCxf
   public SyncServiceClientFactoryCxfRest() {
 
     super();
+  }
+
+  @Override
+  protected ServiceClientTypeHandler getTypeHandler() {
+
+    return ServiceClientTypeHandlerRest.get();
   }
 
   /**
@@ -57,13 +63,9 @@ public class SyncServiceClientFactoryCxfRest extends SyncServiceClientFactoryCxf
   protected <S> S createService(ServiceContext<S> context, String url) {
 
     List<Object> providers = createProviderList(context);
-    return JAXRSClientFactory.create(url, context.getApi(), providers);
-  }
-
-  @Override
-  protected String getServiceTypeFolderName() {
-
-    return ServiceConstants.URL_FOLDER_REST;
+    S service = JAXRSClientFactory.create(url, context.getApi(), providers);
+    applyAspects(context, service);
+    return service;
   }
 
   @Override
@@ -76,12 +78,6 @@ public class SyncServiceClientFactoryCxfRest extends SyncServiceClientFactoryCxf
         webClient.header(headerName, context.getHeader(headerName));
       }
     }
-  }
-
-  @Override
-  protected boolean isResponsibleForService(ServiceContext<?> context) {
-
-    return context.getApi().isAnnotationPresent(Path.class);
   }
 
   /**
