+++ modules/http-client-rest/src/main/java/com/devonfw/module/httpclient/common/impl/rest/AsyncServiceClientFactoryHttpRest.java
@@ -1,34 +1,20 @@
 package com.devonfw.module.httpclient.common.impl.rest;
 
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
 import javax.inject.Inject;
-import javax.ws.rs.Path;
 
-import com.devonfw.module.httpclient.common.impl.AsyncServiceClientFactoryHttp;
-import com.devonfw.module.httpclient.common.impl.ServiceHttpClient;
-import com.devonfw.module.json.common.base.ObjectMapperFactory;
 import com.devonfw.module.service.common.api.client.AsyncServiceClient;
 import com.devonfw.module.service.common.api.client.context.ServiceContext;
 import com.devonfw.module.service.common.api.client.sync.SyncServiceClientFactory;
-import com.devonfw.module.service.common.api.constants.ServiceConstants;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.devonfw.module.service.common.base.client.AbstractAsyncServiceClientFactoryRest;
 
 /**
  * Implementation of {@link SyncServiceClientFactory} for JAX-RS REST service clients using Apache CXF.
  *
  * @since 2020.08.001
  */
-public class AsyncServiceClientFactoryHttpRest extends AsyncServiceClientFactoryHttp {
+public class AsyncServiceClientFactoryHttpRest extends AbstractAsyncServiceClientFactoryRest {
 
-  private final Map<Class<?>, RestServiceMetadata<?>> serviceMetadataMap;
-
-  private ClassLoader classLoader;
-
-  private ObjectMapperFactory objectMapperFactory;
-
-  private ObjectMapper objectMapper;
+  private ServiceClientHttpAdapterRest adapter;
 
   /**
    * The constructor.
@@ -36,87 +22,29 @@ public class AsyncServiceClientFactoryHttpRest extends AsyncServiceClientFactory
   public AsyncServiceClientFactoryHttpRest() {
 
     super();
-    this.serviceMetadataMap = new ConcurrentHashMap<>();
   }
 
   /**
-   * @return objectMapper
+   * @return the {@link ServiceClientHttpAdapterRest}.
    */
-  public ObjectMapper getObjectMapper() {
+  public ServiceClientHttpAdapterRest getAdapter() {
 
-    if (this.objectMapper == null) {
-      this.objectMapper = this.objectMapperFactory.createInstance();
-    }
-    return this.objectMapper;
+    return this.adapter;
   }
 
   /**
-   * @return the {@link ObjectMapperFactory}.
-   */
-  public ObjectMapperFactory getObjectMapperFactory() {
-
-    return this.objectMapperFactory;
-  }
-
-  /**
-   * @param objectMapperFactory the {@link ObjectMapperFactory} to {@link Inject}.
+   * @param adapter new value of {@link #getAdapter()}.
    */
   @Inject
-  public void setObjectMapperFactory(ObjectMapperFactory objectMapperFactory) {
+  public void setAdapter(ServiceClientHttpAdapterRest adapter) {
 
-    this.objectMapperFactory = objectMapperFactory;
-  }
-
-  /**
-   * @return the {@link ClassLoader} to use.
-   */
-  public ClassLoader getClassLoader() {
-
-    if (this.classLoader == null) {
-      this.classLoader = Thread.currentThread().getContextClassLoader();
-    }
-    return this.classLoader;
-  }
-
-  /**
-   * @param classLoader new value of {@link #getClassLoader()}.
-   */
-  public void setClassLoader(ClassLoader classLoader) {
-
-    this.classLoader = classLoader;
+    this.adapter = adapter;
   }
 
   @Override
-  protected String getServiceTypeFolderName() {
+  protected <S> AsyncServiceClient<S> createService(ServiceContext<S> context, String baseUrl) {
 
-    return ServiceConstants.URL_FOLDER_REST;
-  }
-
-  @Override
-  protected boolean isResponsibleForService(ServiceContext<?> context) {
-
-    return context.getApi().isAnnotationPresent(Path.class);
-  }
-
-  /**
-   * @param <S> type of the {@link ServiceContext#getApi() service API}.
-   * @param context the {@link ServiceContext}.
-   * @return the {@link RestServiceMetadata} for the given {@link ServiceContext}.
-   */
-  @SuppressWarnings("unchecked")
-  protected <S> RestServiceMetadata<S> getServiceMetadata(ServiceContext<S> context) {
-
-    Class<S> api = context.getApi();
-    RestServiceMetadata<S> serviceMetadata = (RestServiceMetadata<S>) this.serviceMetadataMap.computeIfAbsent(api,
-        s -> new RestServiceMetadata<>(context));
-    return serviceMetadata;
-  }
-
-  @Override
-  protected <S> AsyncServiceClient<S> createService(ServiceContext<S> context, String url) {
-
-    ServiceHttpClient client = new ServiceHttpClient(getHttpClient(), url);
-    return AsyncServiceHttpClientRest.of(context, client, this);
+    return AsyncServiceHttpClientRest.of(context, this.adapter, baseUrl);
   }
 
 }
