+++ modules/http-client-rest/src/main/java/com/devonfw/module/httpclient/common/impl/rest/AsyncServiceHttpClientRest.java
@@ -1,27 +1,11 @@
 package com.devonfw.module.httpclient.common.impl.rest;
 
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpRequest.BodyPublisher;
-import java.net.http.HttpRequest.BodyPublishers;
-import java.net.http.HttpRequest.Builder;
-import java.net.http.HttpResponse;
-import java.nio.file.Path;
-import java.time.temporal.Temporal;
-
-import javax.ws.rs.core.MediaType;
-
 import com.devonfw.module.httpclient.common.impl.AbstractAsyncServiceHttpClient;
-import com.devonfw.module.httpclient.common.impl.ServiceHttpClient;
+import com.devonfw.module.httpclient.common.impl.ServiceClientHttpAdapter;
 import com.devonfw.module.service.common.api.client.AsyncServiceClient;
-import com.devonfw.module.service.common.api.client.async.ServiceClientInvocation;
 import com.devonfw.module.service.common.api.client.async.ServiceClientStub;
 import com.devonfw.module.service.common.api.client.context.ServiceContext;
 import com.devonfw.module.service.common.base.client.async.ServiceClientStubImpl;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 /**
  * Implementation of {@link AsyncServiceClient} for REST (JAX-RS) using Java HTTP client.
@@ -29,113 +13,34 @@ import com.fasterxml.jackson.databind.ObjectMapper;
  * @param <S> type of the {@link #get() service client}.
  * @since 2020.08.001
  */
-public class AsyncServiceHttpClientRest<S>
-    extends AbstractAsyncServiceHttpClient<S, AsyncServiceClientFactoryHttpRest> {
-
-  private final RestServiceMetadata<S> serviceMetadata;
+public class AsyncServiceHttpClientRest<S> extends AbstractAsyncServiceHttpClient<S> {
 
   /**
    * The constructor.
    *
    * @param proxy the {@link #get() service client}.
    * @param stub the {@link ServiceClientStub}.
-   * @param client the {@link ServiceHttpClient} to use.
-   * @param factory the owning {@link AsyncServiceClientFactoryHttpRest}.
+   * @param adapter the {@link ServiceClientHttpAdapter} to use.
+   * @param baseUrl the discovered and resolved base URL of the service to invoke.
    */
-  public AsyncServiceHttpClientRest(S proxy, ServiceClientStub<S> stub, ServiceHttpClient client,
-      AsyncServiceClientFactoryHttpRest factory) {
+  public AsyncServiceHttpClientRest(S proxy, ServiceClientStub<S> stub, ServiceClientHttpAdapter adapter,
+      String baseUrl) {
 
-    super(proxy, stub, client, factory);
-    this.serviceMetadata = factory.getServiceMetadata(stub.getContext());
-  }
-
-  @Override
-  protected Object createResult(HttpResponse<?> response, ServiceClientInvocation<S> invocation) {
-
-    Object body = response.body();
-    if (body == null) {
-      return null;
-    }
-    Class<?> returnType = invocation.getMethod().getReturnType();
-    if ((returnType == void.class) || (returnType == Void.class)) {
-      return null;
-    }
-    if (body instanceof String) {
-      ObjectMapper objectMapper = this.factory.getObjectMapper();
-      if (CharSequence.class.isAssignableFrom(returnType)) {
-        return body;
-      }
-      try {
-        Object result = objectMapper.readValue((String) body, returnType);
-        return result;
-      } catch (Exception e) {
-        throw new IllegalStateException(e);
-      }
-    } else {
-      return handleUnsupportedBody(body);
-    }
-  }
-
-  @Override
-  protected HttpRequest createRequest(ServiceClientInvocation<S> invocation) {
-
-    RestMethodMetadata method = this.serviceMetadata.getMethod(invocation.getMethod());
-    String url = method.getPath().resolve(this.client.getBaseUrl(), invocation);
-    String contentType = null;
-    BodyPublisher body = null;
-    RestParameter bodyParameter = method.getParameters().getBodyParameter();
-    if (bodyParameter != null) {
-      Object value = invocation.getParameter(bodyParameter.index);
-      if (value != null) {
-        if ((value instanceof CharSequence) || (value instanceof Number) || (value instanceof Temporal)) {
-          body = BodyPublishers.ofString(value.toString());
-          // contentType = MediaType.TEXT_PLAIN;
-        } else if (value instanceof File) {
-          body = createBody(((File) value).toPath());
-          contentType = MediaType.APPLICATION_OCTET_STREAM;
-        } else if (value instanceof Path) {
-          body = createBody((Path) value);
-          contentType = MediaType.APPLICATION_OCTET_STREAM;
-        } else {
-          ObjectMapper objectMapper = this.factory.getObjectMapper();
-          try {
-            String json = objectMapper.writeValueAsString(value);
-            body = BodyPublishers.ofString(json);
-            contentType = MediaType.APPLICATION_JSON;
-          } catch (JsonProcessingException e) {
-            throw new IllegalStateException(e);
-          }
-        }
-      }
-    }
-    Builder requestBuilder = this.client.requestBuilder(url, invocation, method.getHttpMethod(), body);
-    if (contentType != null) {
-      requestBuilder.header("Content-Type", contentType);
-    }
-    return requestBuilder.build();
-  }
-
-  private BodyPublisher createBody(Path path) {
-
-    try {
-      return BodyPublishers.ofFile(path);
-    } catch (FileNotFoundException e) {
-      throw new IllegalArgumentException(e);
-    }
+    super(proxy, stub, adapter, baseUrl);
   }
 
   /**
    * @param <S> type of the {@link #get() service client}.
    * @param context the {@link ServiceContext}.
-   * @param client the {@link HttpClient} to use.
-   * @param factory the {@link AsyncServiceClientFactoryHttpRest}.
+   * @param adapter the {@link ServiceClientHttpAdapter} to use.
+   * @param baseUrl the discovered and resolved base URL of the service to invoke.
    * @return the new {@link AsyncServiceHttpClientRest}.
    */
-  public static <S> AsyncServiceHttpClientRest<S> of(ServiceContext<S> context, ServiceHttpClient client,
-      AsyncServiceClientFactoryHttpRest factory) {
+  public static <S> AsyncServiceHttpClientRest<S> of(ServiceContext<S> context, ServiceClientHttpAdapter adapter,
+      String baseUrl) {
 
-    ServiceClientStub<S> stub = ServiceClientStubImpl.of(context, factory.getClassLoader());
-    return new AsyncServiceHttpClientRest<>(stub.getProxy(), stub, client, factory);
+    ServiceClientStub<S> stub = ServiceClientStubImpl.of(context, adapter.getClassLoader());
+    return new AsyncServiceHttpClientRest<>(stub.getProxy(), stub, adapter, baseUrl);
   }
 
 }
