+++ modules/service/pom.xml
@@ -48,10 +48,16 @@
       <optional>true</optional>
     </dependency>
     <dependency>
+      <groupId>jakarta.ws.rs</groupId>
+      <artifactId>jakarta.ws.rs-api</artifactId>
+      <optional>true</optional>
+    </dependency>
+    <dependency>
       <groupId>javax.annotation</groupId>
       <artifactId>javax.annotation-api</artifactId>
       <scope>test</scope>
     </dependency>
+
   </dependencies>
 
 </project>
