+++ modules/service/src/main/java/com/devonfw/module/service/common/api/constants/ServiceConstants.java
@@ -61,4 +61,12 @@ public class ServiceConstants {
    * The variable that resolves to type of the service (e.g. "rest" for REST service and "ws" for SOAP service).
    */
   public static final String VARIABLE_TYPE = "${type}";
+
+  /**
+   * URL suffix for convention to retrieve WSDL from SOAP service.
+   *
+   * @since 2021.04.003
+   */
+  public static final String WSDL_SUFFIX = "?wsdl";
+
 }
