+++ modules/service/src/main/java/com/devonfw/module/service/common/base/client/AbstractAsyncServiceClientFactoryRest.java
@@ -1,0 +1,16 @@
+package com.devonfw.module.service.common.base.client;
+
+/**
+ * Abstract base implementation of {@link AbstractAsyncServiceClientFactory} for REST.
+ *
+ * @since 2021.04.003
+ */
+public abstract class AbstractAsyncServiceClientFactoryRest extends AbstractAsyncServiceClientFactory {
+
+  @Override
+  protected ServiceClientTypeHandler getTypeHandler() {
+
+    return ServiceClientTypeHandlerRest.get();
+  }
+
+}
